+++ pkg/resource/promtail-config/promtail-config.go
@@ -132,6 +132,13 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
     target_label: 'systemd_unit'
   - source_labels: ['__journal__hostname']
     target_label: 'node_name'
+  pipeline_stages:
+  - json:
+      expressions:
+        SYSLOG_IDENTIFIER: SYSLOG_IDENTIFIER
+  - drop:
+      source: SYSLOG_IDENTIFIER
+      value: audit
 - job_name: systemd_journal_var
   journal:
     path: /var/log/journal
@@ -144,6 +151,13 @@ func GeneratePromtailConfig(lc loggedcluster.Interface) (v1.ConfigMap, error) {
     target_label: 'systemd_unit'
   - source_labels: ['__journal__hostname']
     target_label: 'node_name'
+  pipeline_stages:
+  - json:
+      expressions:
+        SYSLOG_IDENTIFIER: SYSLOG_IDENTIFIER
+  - drop:
+      source: SYSLOG_IDENTIFIER
+      value: audit
 - job_name: kubernetes-audit
   static_configs:
   - targets: